transactions: refactor search mocks location
